Setup pip package
Set the library up as a pip package that installs  afmplot and afmfit console scripts that correspond to plot_datashop and process_datashop. Also cleaned up the way that corss validation is done in both models to correspond to modern scikit conventions.
