+++ .gitignore
@@ -1,3 +1,4 @@
 # ignore precompiled stuff
 __pycache__/
 *.py[cod]
+*.egg-info
