+++ pyafm/util.py
